Cleanup statue movement
