--- Assets/Scenes/Nav Testing.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 24095b29ef47fdd4ba940309dabda453
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
