--- Assets/Scenes/Nav Testing.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 34b44fe6ce833e94583564914fd5bd9c
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
