--- Assets/Scenes/Nav Testing/NavMesh-Floor.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5a3928ead82966b47869452d35a8971d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
