--- Assets/Scripts/Enemies/Enemy Pathing.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2ac6a9ccf55e6a498b6d66609ea9db6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
